refactor: split map state up into providers
