+++ src/app/(private)/map/[id]/components/lists/MarkerList.tsx
@@ -1,6 +1,8 @@
 import { Check, Database, Pencil, Table, Trash2 } from "lucide-react";
 import { useContext, useState } from "react";
 import { MapContext } from "@/app/(private)/map/[id]/context/MapContext";
+import { MarkerAndTurfContext } from "@/app/(private)/map/[id]/context/MarkerAndTurfContext";
+import { TableContext } from "@/app/(private)/map/[id]/context/TableContext";
 import { Button } from "@/shadcn/ui/button";
 import {
   ContextMenu,
@@ -13,17 +15,18 @@ import DataSourceIcon from "../DataSourceIcon";
 import Loading from "../Loading";
 
 export default function MarkerList() {
+  const { mapRef, dataSourcesQuery, viewConfig } = useContext(MapContext);
+
   const {
-    mapRef,
-    dataSourcesQuery,
-    viewConfig,
     placedMarkers,
     placedMarkersLoading,
-    selectedDataSourceId,
-    handleDataSourceSelect,
     deletePlacedMarker,
     updatePlacedMarker,
-  } = useContext(MapContext);
+  } = useContext(MarkerAndTurfContext);
+
+  const { selectedDataSourceId, handleDataSourceSelect } =
+    useContext(TableContext);
+
   const [editingIndex, setEditingIndex] = useState<number | null>(null);
   const [editText, setEditText] = useState("");
   const [contextMenuIndex, setContextMenuIndex] = useState<number | null>(null);
