+++ src/app/(private)/map/[id]/context/MapContext.tsx
@@ -3,23 +3,15 @@ import { RefObject, createContext } from "react";
 import { MapRef } from "react-map-gl/mapbox";
 import {
   AreaSetGroupCode,
-  AreaStatsQuery,
-  AreaStatsQueryVariables,
   BoundingBoxInput,
-  DataRecordsQuery,
-  DataRecordsQueryVariables,
   DataSourcesQuery,
+  MapQuery,
+  MapQueryVariables,
   MapStyleName,
   MapViewConfigInput,
-  PlacedMarker,
-  SortInput,
-  Turf,
 } from "@/__generated__/types";
 import { DEFAULT_ZOOM } from "@/constants";
-import { MarkerData } from "@/types";
-import { ChoroplethLayerConfig, getChoroplethLayerConfig } from "../sources";
 import mapStyles from "../styles";
-import { MarkerQueriesResult } from "../types";
 
 export class ViewConfig implements MapViewConfigInput {
   public areaDataSourceId = "";
@@ -64,39 +56,8 @@ export const MapContext = createContext<{
   mapRef: RefObject<MapRef | null> | null;
 
   /* State */
-  boundariesPanelOpen: boolean;
-  setBoundariesPanelOpen: (open: boolean) => void;
-
   boundingBox: BoundingBoxInput | null;
   setBoundingBox: (boundingBox: BoundingBoxInput | null) => void;
-
-  editingTurf: Turf | null;
-  setEditingTurf: (turf: Turf | null) => void;
-
-  placedMarkers: PlacedMarker[];
-  placedMarkersLoading: boolean;
-  deletePlacedMarker: (id: string) => void;
-  insertPlacedMarker: (placedMarker: PlacedMarker) => void;
-  updatePlacedMarker: (placedMarker: PlacedMarker) => void;
-
-  selectedDataSourceId: string;
-  handleDataSourceSelect: (dataSourceId: string) => void;
-
-  selectedMarker: MarkerData | null;
-  setSelectedMarker: (marker: MarkerData | null) => void;
-
-  tableFilter: string;
-  setTableFilter: (filter: string) => void;
-  tablePage: number;
-  setTablePage: (page: number) => void;
-  tableSort: SortInput[];
-  setTableSort: (tableSort: SortInput[]) => void;
-
-  turfs: Turf[];
-  turfsLoading: boolean;
-  deleteTurf: (id: string) => void;
-  insertTurf: (turf: Turf) => void;
-  updateTurf: (turf: Turf) => void;
 
   viewConfig: ViewConfig;
   updateViewConfig: (config: Partial<ViewConfig>) => void;
@@ -108,60 +69,21 @@ export const MapContext = createContext<{
   setZoom: (zoom: number) => void;
 
   /* GraphQL Queries */
-  areaStatsQuery: QueryResult<AreaStatsQuery, AreaStatsQueryVariables> | null;
-  dataRecordsQuery: QueryResult<
-    DataRecordsQuery,
-    DataRecordsQueryVariables
-  > | null;
   dataSourcesQuery: QueryResult<DataSourcesQuery> | null;
-  markerQueries: MarkerQueriesResult | null;
-
-  /* Derived Properties */
-  choroplethLayerConfig: ChoroplethLayerConfig;
-  selectedRecordId: string | null;
-  setSelectedRecordId: (recordId: string | null) => void;
+  mapQuery: QueryResult<MapQuery, MapQueryVariables> | null;
 }>({
   mapId: null,
   mapName: null,
   setMapName: () => null,
   mapRef: null,
-  boundariesPanelOpen: false,
-  setBoundariesPanelOpen: () => null,
   boundingBox: null,
   setBoundingBox: () => null,
-  choroplethLayerConfig: getChoroplethLayerConfig(null, DEFAULT_ZOOM),
-  editingTurf: null,
-  setEditingTurf: () => null,
-  placedMarkers: [],
-  placedMarkersLoading: false,
-  deletePlacedMarker: () => null,
-  insertPlacedMarker: () => null,
-  updatePlacedMarker: () => null,
-  selectedMarker: null,
-  setSelectedMarker: () => null,
-  tableFilter: "",
-  setTableFilter: () => null,
-  tablePage: 0,
-  setTablePage: () => null,
-  tableSort: [],
-  setTableSort: () => null,
-  turfs: [],
-  turfsLoading: false,
-  deleteTurf: () => null,
-  insertTurf: () => null,
-  updateTurf: () => null,
   viewConfig: new ViewConfig(),
   updateViewConfig: () => null,
   viewId: null,
   setViewId: () => null,
   zoom: DEFAULT_ZOOM,
   setZoom: () => null,
-  selectedDataSourceId: "",
-  handleDataSourceSelect: () => null,
-  selectedRecordId: null,
-  setSelectedRecordId: () => null,
-  areaStatsQuery: null,
-  dataRecordsQuery: null,
   dataSourcesQuery: null,
-  markerQueries: null,
+  mapQuery: null,
 });
