+++ src/app/(private)/map/[id]/page.tsx
@@ -1,4 +1,8 @@
 import MapPage from "./MapPage";
+import ChoroplethProvider from "./providers/ChoroplethProvider";
+import MapProvider from "./providers/MapProvider";
+import MarkerAndTurfProvider from "./providers/MarkerAndTurfProvider";
+import TableProvider from "./providers/TableProvider";
 
 export default async function MapPageWrapper({
   params,
@@ -6,5 +10,15 @@ export default async function MapPageWrapper({
   params: Promise<{ id: string }>;
 }) {
   const { id } = await params;
-  return <MapPage mapId={id} />;
+  return (
+    <MapProvider mapId={id}>
+      <ChoroplethProvider>
+        <MarkerAndTurfProvider>
+          <TableProvider>
+            <MapPage />
+          </TableProvider>
+        </MarkerAndTurfProvider>
+      </ChoroplethProvider>
+    </MapProvider>
+  );
 }
